pyproject.toml
[project]
name = "daily-ai-insight-engine"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "jinja2>=3.1.6",
    "matplotlib>=3.10.8",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sniffio>=1.3.1",
    "wordcloud>=1.9.6",
    "zhipuai>=2.1.5.20250825",
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "schedule>=1.2.0",
]
